structured the folder
