+++ index.html
@@ -9,6 +9,6 @@
   <body>
     <div id="root"></div>
 
-    <script type="module" src="./App.js"></script>
+    <script type="module" src="./src/App.js"></script>
   </body>
 </html>
